gemini_service/requirements.txt
# GeminiService Dependencies
# LLM operations and HTTP client to StoringService

fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.1
httpx==0.25.1
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0

